+++ ipatests/pytest_plugins/integration/config.py
@@ -88,11 +88,11 @@ class Config(pytest_multihost.config.Config):
 
     @classmethod
     def from_env(cls, env):
-        from ipatests.test_integration.env_config import config_from_env
+        from ipatests.pytest_plugins.integration.env_config import config_from_env
         return config_from_env(env)
 
     def to_env(self, **kwargs):
-        from ipatests.test_integration.env_config import config_to_env
+        from ipatests.pytest_plugins.integration.env_config import config_to_env
         return config_to_env(self, **kwargs)
 
 
@@ -155,9 +155,9 @@ class Domain(pytest_multihost.config.Domain):
 
     @classmethod
     def from_env(cls, env, config, index, domain_type):
-        from ipatests.test_integration.env_config import domain_from_env
+        from ipatests.pytest_plugins.integration.env_config import domain_from_env
         return domain_from_env(env, config, index, domain_type)
 
     def to_env(self, **kwargs):
-        from ipatests.test_integration.env_config import domain_to_env
+        from ipatests.pytest_plugins.integration.env_config import domain_to_env
         return domain_to_env(self, **kwargs)
