+++ ipatests/test_integration/tasks.py
@@ -37,7 +37,7 @@ from ipaplatform.paths import paths
 from ipapython.dn import DN
 from ipapython.ipa_log_manager import log_mgr
 from ipatests.test_integration import util
-from ipatests.test_integration.env_config import env_to_script
+from ipatests.pytest_plugins.integration.env_config import env_to_script
 from ipatests.test_integration.host import Host
 from ipalib import errors
 from ipalib.util import get_reverse_zone_default, verify_host_resolvable
