+++ ipatests/test_integration/test_idviews.py
@@ -7,7 +7,7 @@ import re
 import string
 from ipatests.test_integration import tasks
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration.env_config import get_global_config
+from ipatests.pytest_plugins.integration.env_config import get_global_config
 from ipaplatform.paths import paths
 config = get_global_config()
