+++ ipatests/test_integration/test_topology.py
@@ -8,7 +8,7 @@ import pytest
 
 from ipatests.test_integration.base import IntegrationTest
 from ipatests.test_integration import tasks
-from ipatests.test_integration.env_config import get_global_config
+from ipatests.pytest_plugins.integration.env_config import get_global_config
 from ipalib.constants import DOMAIN_SUFFIX_NAME
 from ipatests.util import assert_deepequal
